fix bugs: - use int for planet id - set planet id on updates

+++ zoom_navigator_memory_access/src/main/java/info/nanodesu/reader/FullUnitInfo.java
@@ -7,15 +7,15 @@ public class FullUnitInfo {
 	private float x;
 	private float y;
 	private float z;
-	private float planetId;
+	private int planetId;
 	private float currentHp;
 	private float maxHp;
 		
-	public float getPlanetId() {
+	public int getPlanetId() {
 		return planetId;
 	}
 	
-	public void setPlanetId(float planetId) {
+	public void setPlanetId(int planetId) {
 		this.planetId = planetId;
 	}
